devex 1218: upgraded from v2 to v3; update some calls; read from source table and write to destination table instead of deploy table
