+++ .circleci/config.yml
@@ -293,7 +293,7 @@ jobs:
           name: Track successful migrations
           command: |
             if [ "$MIGRATE_FLAG" == "true" ]; then
-              node ./web-api/track-successful-migrations.js
+              npx ts-node --transpile-only ./web-api/track-successful-migrations.ts
             fi
       - run:
           name: Enable Destination Table Stream
