+++ docs/backend.md
@@ -93,7 +93,7 @@ Before diving into the technical aspects of the backend, it might be useful just
 │   │   │       ├── main.tf // the main terraform entrypoint
 ├── swagger.json // the swagger .json we update when api endpoints are changed or added
 ├── switch-cognito-triggers-color.js // used to switch the color of the cognito triggers
-├── track-successful-migrations.js // tracks which migration scripts have run after doing a migration
+├── track-successful-migrations.ts // tracks which migration scripts have run after doing a migration
 ├── verify-ses-email.sh // used to verify SES 
 ```
