+++ docs/migrations.md
@@ -113,7 +113,7 @@ If learning via a video if more your style, we have a short recording explaining
 
 ## Keeping Track of Migrations
 
-When our CI/CD process runs the migration process, it keeps track of which migration scripts have ran inside our `efcms-deploy-$ENV` table using the `./web-api/track-successful-migrations.js` node script.  This script looks at the migration directory and writes entries into dynamo to track the migrations that have ran.  An example of one of these records looks like this:
+When our CI/CD process runs the migration process, it keeps track of which migration scripts have ran inside our `efcms-deploy-$ENV` table using the `./web-api/track-successful-migrations.ts` node script.  This script looks at the migration directory and writes entries into dynamo to track the migrations that have ran.  An example of one of these records looks like this:
 
 ```javascript
 {
@@ -193,7 +193,7 @@ To help further understand these terms, take a look at the following diagram whi
 
 ![Blue-Green Migration Architecture](https://user-images.githubusercontent.com/1868782/117465361-9f83e400-af1f-11eb-8844-b14fefa2c3d2.png)
 
-As an overview of this diagram, when we do a deploy in circle, part of the CI/CD pipeline will check using the `./web-api/is-migration-needed.js` script if we need to run a new migration script.  If a new migration is needed, CI/CD will run our  `npm run deploy:migration` script which will run terraform and create some of the resources displayed in this diagram above.  For example, this terraform deploy will create the SQS queues, the dead letter queues, the migration lambda, the live migration streams, and the segment migration lambda, etc.  
+As an overview of this diagram, when we do a deploy in circle, part of the CI/CD pipeline will check using the `./web-api/is-migration-needed.ts` script if we need to run a new migration script.  If a new migration is needed, CI/CD will run our  `npm run deploy:migration` script which will run terraform and create some of the resources displayed in this diagram above.  For example, this terraform deploy will create the SQS queues, the dead letter queues, the migration lambda, the live migration streams, and the segment migration lambda, etc.  
 
 ?> Please checkout our [Dawson's Terraform documentation](terraform?id=migration-terraform) for more information about terraform and our project.
